Move circuits state management logic to its own file
