+++ saplings/circuits/src/components/Content.js
@@ -14,78 +14,15 @@
  * limitations under the License.
  */
 
-import React, { useReducer } from 'react';
+import React from 'react';
 import { useLocalNodeState } from '../state/localNode';
-import mockCircuits from '../mockData/mockCircuits';
-import mockProposals from '../mockData/mockProposals';
-import { processCircuits } from '../data/processCircuits';
+import { useCircuitsState } from '../state/circuits';
 
 import './Content.scss';
 
-const circuitsReducer = (state, action) => {
-  switch (action.type) {
-    case 'sort': {
-      const sortedCircuits = action.sortCircuits(
-        state.filteredCircuits,
-        action.sort
-      );
-      return { ...state, filteredCircuits: sortedCircuits };
-    }
-    case 'filter': {
-      const filteredCircuits = action.filterCircuits(
-        state.circuits,
-        action.filter
-      );
-      return { ...state, filteredCircuits };
-    }
-    default:
-      throw new Error(`unhandled action type: ${action.type}`);
-  }
-};
-
-const filterCircuits = (circuits, filterBy) => {
-  if (filterBy.filterTerm.length === 0) {
-    return circuits;
-  }
-  const filteredCircuits = circuits.filter(circuit => {
-    if (circuit.id.toLowerCase().indexOf(filterBy.filterTerm) > -1) {
-      return true;
-    }
-    if (
-      circuit.managementType.toLowerCase().indexOf(filterBy.filterTerm) > -1
-    ) {
-      return true;
-    }
-    if (circuit.comments.toLowerCase().indexOf(filterBy.filterTerm) > -1) {
-      return true;
-    }
-    if (
-      circuit.members.filter(
-        member => member.toLowerCase().indexOf(filterBy.filterTerm) > -1
-      ).length > 0
-    ) {
-      return true;
-    }
-    if (
-      circuit.roster.filter(
-        service => service.service_type.indexOf(filterBy.filterTerm) > -1
-      ).length > 0
-    ) {
-      return true;
-    }
-    return false;
-  });
-
-  return filteredCircuits;
-};
-
 const Content = () => {
-  const circuits = processCircuits(mockCircuits.concat(mockProposals));
+  const [circuitState, circuitsDispatch] = useCircuitsState();
 
-  const [circuitState, circuitsDispatch] = useReducer(circuitsReducer, {
-    circuits,
-    filteredCircuits: circuits
-  });
   const nodeID = useLocalNodeState();
   const totalCircuits = circuitState.circuits.length;
   let actionRequired = 0;
@@ -117,7 +54,6 @@ const Content = () => {
           onKeyUp={event => {
             circuitsDispatch({
               type: 'filter',
-              filterCircuits,
               filter: {
                 filterTerm: event.target.value.toLowerCase()
               }
